third_party/blink/web_tests/fast/mediastream/constructors.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests the JSEP PeerConnection related constructors.");

shouldBeTrue("typeof RTCPeerConnection === 'function'");
shouldBeTrue("typeof RTCSessionDescription === 'function'");
shouldBeTrue("typeof RTCIceCandidate === 'function'");

shouldThrow("RTCPeerConnection()");
shouldThrow("RTCSessionDescription()");
shouldThrow("RTCIceCandidate()");

shouldNotThrow("new RTCPeerConnection({iceServers:[{url:'stun:foobar.com:12345'}]}, null);");
shouldNotThrow("new RTCSessionDescription({type:'offer',sdp:'foobar'});");

window.jsTestIsAsync = false;
</script>
</body>
</html>
